requirements.txt
Flask
vk_api
gspread
google-auth
